backend/requirements.txt
fastapi==0.115.6
python-multipart==0.0.20
uvicorn[standard]==0.34.0
httpx==0.28.1
mcp==1.0.0
pydantic==2.10.4
python-dotenv==1.0.1
